added hooks for uploadFile
