+++ package.json
@@ -25,6 +25,7 @@
         "@as-integrations/next": "^3.0.0",
         "@auth/mongodb-adapter": "^2.0.12",
         "@aws-sdk/client-s3": "^3.569.0",
+        "@aws-sdk/s3-request-presigner": "^3.645.0",
         "@googlemaps/js-api-loader": "^1.16.6",
         "@googlemaps/routing": "^1.1.0",
         "@vis.gl/react-google-maps": "^1.1.0",
