+++ src/app/api/upload/route.ts
@@ -2,6 +2,7 @@ import { NextRequest, NextResponse } from "next/server";
 import { ObjectCannedACL, PutObjectCommand } from "@aws-sdk/client-s3";
 import { customAlphabet } from "nanoid";
 import { s3Client } from "@/app/api/upload/s3client";
+import { getSignedUrl, S3RequestPresigner } from "@aws-sdk/s3-request-presigner";
 export const POST = async (req: NextRequest) => {
     const data = await req.formData();
     const file = data.get("file");
@@ -10,8 +11,6 @@ export const POST = async (req: NextRequest) => {
     if (!(file instanceof File) || !validImageTypes.includes(file.type)) {
         return NextResponse.json("Invalid file type", { status: 400 });
     }
-
-    console.log(file);
 
     const bytes = await file.arrayBuffer();
     const buffer = Buffer.from(bytes);
@@ -25,11 +24,19 @@ export const POST = async (req: NextRequest) => {
         ACL: ObjectCannedACL.public_read,
     };
 
+    const fileUrl = `${process.env.DO_SPACES_URL}/${bucketParams.Bucket}/${bucketParams.Key}`;
+    //Получаем ссылку до загрузки файла в бд
+    const presignUrl = await getSignedUrl(s3Client, new PutObjectCommand(bucketParams), { expiresIn: 3600 });
+
+    //Загружаем файл в бд
     const result = await s3Client.send(new PutObjectCommand(bucketParams));
-    console.log(result);
+
+    //console.log(presignUrl);
 
     return NextResponse.json({
         id: result.ETag,
-        url: `${process.env.DO_SPACES_URL}/${bucketParams.Bucket}/${bucketParams.Key}`,
+        //url: `${process.env.DO_SPACES_URL}/${bucketParams.Bucket}/${bucketParams.Key}`,
+        uploadUrl: presignUrl,
+        fileUrl: fileUrl,
     });
 };
