+++ src/components/widgets/UploadFile/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useUploadFile } from "./useUploadFile";
